using int64_t

+++ include/one.hpp
@@ -9,20 +9,20 @@ namespace aoc
 
 static std::array<char, 2> intermediate = { 'X', 'X' };
 
-class One final : public Day<int>
+class One final : public Day<int64_t>
 {
   public:
     One ();
 
     ~One () final;
 
-    int one () final;
+    int64_t one () final;
 
-    int two () final;
+    int64_t two () final;
 
-    static int calibrate (const std::string &input);
+    static int64_t calibrate (const std::string &input);
 
-    static int recalibrate (const std::string &input);
+    static int64_t recalibrate (const std::string &input);
 };
 } // aoc
